Bug fixes
Due to the fact that it often takes a while for the kernel to detect disks in real world scenarios, I have added a wait before scanning for disks to boot from. This wait is set at two seconds, which is more than enough. On high performance solid state devices, this wait is not needed and can be removed.